Moved most code from FetchPrebuiltUE4 to a class library
