+++ FetchPrebuiltUE4.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DistributionTools", "Distri
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FetchPrebuiltUE4", "FetchPrebuiltUE4\FetchPrebuiltUE4.csproj", "{3C8C5A9B-343A-454F-B06C-41E4EB2FC332}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "FetchPrebuiltUE4Lib", "FetchPrebuiltUE4Lib\FetchPrebuiltUE4Lib.csproj", "{2C0C7A1D-7B45-4115-AE54-7F3A4630BA3A}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{3C8C5A9B-343A-454F-B06C-41E4EB2FC332}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{3C8C5A9B-343A-454F-B06C-41E4EB2FC332}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{3C8C5A9B-343A-454F-B06C-41E4EB2FC332}.Release|Any CPU.Build.0 = Release|Any CPU
+		{2C0C7A1D-7B45-4115-AE54-7F3A4630BA3A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{2C0C7A1D-7B45-4115-AE54-7F3A4630BA3A}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{2C0C7A1D-7B45-4115-AE54-7F3A4630BA3A}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{2C0C7A1D-7B45-4115-AE54-7F3A4630BA3A}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
